chore(card): rename context file and restructure internal directory
