+++ src/components/card/card-row/card-row.spec.tsx
@@ -7,7 +7,7 @@ import {
 } from "../../../__spec_helper__/__internal__/test-utils";
 import { rootTagTest } from "../../../__internal__/utils/helpers/tags/tags-specs";
 import StyledCardRow from "./card-row.style";
-import CardContext, { CardContextProps } from "../__internal__/card-context";
+import CardContext, { CardContextProps } from "../__internal__/card.context";
 
 describe("CardRow", () => {
   it("renders children correctly", () => {
