+++ src/components/card/card.style.ts
@@ -2,7 +2,7 @@ import styled, { css } from "styled-components";
 import { margin, MarginProps } from "styled-system";
 import * as DesignTokens from "@sage/design-tokens/js/base/common";
 import baseTheme from "../../style/themes/base";
-import { CardContextProps } from "./__internal__/card-context";
+import { CardContextProps } from "./__internal__/card.context";
 import { CardProps } from "./card.component";
 import addFocusStyling from "../../style/utils/add-focus-styling";
